v3.1.14

+++ database/migrations/2019_06_18_135639_update_custom_view_summary.php
@@ -28,7 +28,9 @@ class UpdateCustomViewSummary extends Migration
     public function down()
     {
         Schema::table('custom_view_summaries', function($table) {
-            $table->dropColumn('options');
+            if (Schema::hasColumn('custom_view_summaries', 'options')) {
+                $table->dropColumn('options');
+            }
         });
     }
 }
